:recycle: Move to ast_structure_convert convert function
